# openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Uganda-first professional + dating network. Envelope: `{code:1|0, message, data}`. Auth: Bearer JWT.",
    "title": "LinkUp API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/interests/me": {
      "get": {
        "operationId": "get_v1_interest.my_interests",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/interests/me",
        "tags": [
          "interests"
        ]
      },
      "post": {
        "operationId": "post_v1_interest.set_interests",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /v1/interests/me",
        "tags": [
          "interests"
        ]
      }
    },
    "/v1/interests/me/{tag_id}": {
      "delete": {
        "operationId": "delete_v1_interest.remove_interest",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "DELETE /v1/interests/me/{tag_id}",
        "tags": [
          "interests"
        ]
      },
      "put": {
        "operationId": "put_v1_interest.update_interest",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "PUT /v1/interests/me/{tag_id}",
        "tags": [
          "interests"
        ]
      }
    },
    "/v1/interests/search": {
      "get": {
        "operationId": "get_v1_interest.search",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/interests/search",
        "tags": [
          "interests"
        ]
      }
    },
    "/v1/interests/signal": {
      "post": {
        "operationId": "post_v1_interest.signal_interest",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /v1/interests/signal",
        "tags": [
          "interests"
        ]
      }
    },
    "/v1/interests/suggestions": {
      "get": {
        "operationId": "get_v1_interest.suggestions",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/interests/suggestions",
        "tags": [
          "interests"
        ]
      }
    },
    "/v1/interests/taxonomy": {
      "get": {
        "operationId": "get_v1_interest.taxonomy",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/interests/taxonomy",
        "tags": [
          "interests"
        ]
      }
    },
...
    "/v1/links/suggestions": {
      "get": {
        "operationId": "get_v1_links.suggestions",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/links/suggestions",
        "tags": [
          "links"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
...
}